Use generalized test framework for testing
Also moves the remote file fetching to clawutil
