+++ examples/tsunami/chile2010/maketopo.py
@@ -8,7 +8,9 @@ but now they are explicit below.
 Call functions with makeplots==True to create plots of topo, slip, and dtopo.
 """
 
-import os,sys
+import os
+
+import clawpack.clawutil.data
 
 try:
     CLAW = os.environ['CLAW']
@@ -22,11 +24,11 @@ def get_topo(makeplots=False):
     """
     Retrieve the topo file from the GeoClaw repository.
     """
-    from clawpack.geoclaw import topotools, util
+    from clawpack.geoclaw import topotools
     topo_fname = 'etopo10min120W60W60S0S.asc'
     url = 'http://www.geoclaw.org/topo/etopo/' + topo_fname
-    util.get_remote_file(url, output_dir=scratch_dir, file_name=topo_fname,
-            verbose=True)
+    clawpack.clawutil.data.get_remote_file(url, output_dir=scratch_dir, 
+            file_name=topo_fname, verbose=True)
 
     if makeplots:
         from matplotlib import pyplot as plt
